@@ -14,45 +14,36 @@
     },
 
     "require": {
-        "php": ">=8.0",
-        "symfony/flex": "^1.17",
-        "symfony/framework-bundle": "5.4.*",
-        "symfony/runtime": "5.4.*",
-        "symfony/security-bundle": "5.4.*",
-        "doctrine/orm": "^2.10",
-        "doctrine/doctrine-bundle": "^2.5",
-        "symfony/mailer": "5.4.*",
-        "symfony/monolog-bundle": "^3.7",
-        "symfony/validator": "5.4.*"
+        "php": ">=8.2",
+        "symfony/flex": "^2.4",
+        "symfony/framework-bundle": "7.0.*",
+        "symfony/runtime": "7.0.*",
+        "symfony/security-bundle": "7.0.*",
+        "doctrine/orm": "^2.17",
+        "symfony/mailer": "7.0.*",
+        "symfony/validator": "7.0.*",
+        "symfony/twig-bundle": "7.0.*",
+        "symfony/yaml": "7.0.*",
+        "symfony/asset": "7.0.*",
+        "symfony/form": "7.0.*",
+        "twig/extra-bundle": "^2.12|^3.0",
+        "twig/twig": "^2.12|^3.0",
+        "symfony/monolog-bundle": "^3.10",
+        "doctrine/doctrine-bundle": "^2.14",
+        "symfony/dotenv": "7.0.*"
     },
     "config": {
         "platform": {
-            "php": "8.0.0"
+            "php": "8.2.0"
         },
         "allow-plugins": {
             "symfony/flex": true,
             "symfony/runtime": true
         }
     },
-
-    "scripts": {
-        "symfony-scripts": [
-            "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
-        ],
-        "post-install-cmd": [
-            "@symfony-scripts"
-        ],
-        "post-update-cmd": [
-            "@symfony-scripts"
-        ],
-        "auto-scripts": {
-            "cache:clear": "symfony-cmd",
-            "assets:install %PUBLIC_DIR%": "symfony-cmd"
-        }
+    "require-dev": {
+        "symfony/web-profiler-bundle": "^7.2",
+        "symfony/debug-bundle": "^7.2",
+        "symfony/maker-bundle": "^1.62"
     }
 }
